adding Global Exception Handling

+++ src/main/java/com/caioLessa/dscommerce/services/ProductService.java
@@ -3,10 +3,13 @@ package com.caioLessa.dscommerce.services;
 import com.caioLessa.dscommerce.dto.ProductDTO;
 import com.caioLessa.dscommerce.entities.Product;
 import com.caioLessa.dscommerce.repositories.ProductRepository;
+import com.caioLessa.dscommerce.services.exceptions.DatabaseException;
 import com.caioLessa.dscommerce.services.exceptions.ResourceNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -19,7 +22,7 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado para o ID informado!"));
+        Product product = productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado para o ID " + id));
         return new ProductDTO(product);
     }
 
@@ -45,9 +48,16 @@ public class ProductService {
         return new ProductDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if(!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("\"Produto não encontrado para o ID " + id);
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch(DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial!");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
